requirements.txt
obs-websocket-py